Timestamp improvements on data and meta-data.
Timestamps on hits records is now being recorded in the FITS files.
This timestamp is system time to nanosecond accuracy, expressed
as decimal seconds.  That the system uses NTP is assumed.

Timestamps on meta-data is changed from a two field expression to
decimal seconds to match the timestamp format in the hits records.